feat: re-enable shadow-dom & add width and height attrs
This commit re-enables the shadow dom, imports the video.js styles in a
`<link>` tag, and creates the `width` and `height` attributes on the
`ix-video` WC.

The shadow dom can be re-enable since the link tag is inside the
shadow-dom, scoping the CSS styles correctly.

The `width` and `height` attributes are needed in order to 1) enable
customization of the component dimensions and 2) enable the component to
take the full width and height of its container if its w/h attributes
are not set.

Finally, this commit changes the `videoSrc` attribute to be `source`.