feat: disable shadow dom and remove link tag
Before this commit, there was a flash-of-unstyled-content (FOUC) that
could occur on component first render.

After this commit, the FOUC no longer occurs.

This was achieved by removing the shadow dom and the inline `<link>`
tag. Unfrotnately this means that some of the benefits of WC (shadow-dom
scoping styles) are also removed. But due to the video.js API we've as
of yet found a good way around this issue.