feat: define create-style-defaults fn
This commit refactors some of the type casting done in the helper file.
It also defines the `createStyleDefaults()` function. This function will
return an object with some style mappings that our component will
default to. This is necessary because:

1. We cannot assume the user will pass win values for `width/height`
2. We cannot assume that `dataSetup` won't have it's own `width/height`
   values defined in the case the user didn't set the attributes
3. In the event no w/h values are defined, we should take up 100% of the
   video container

For some reason, although that's the default video.js behavior, we have
to explicity add this style to our player or we won't take up 100% of
available space when no w/h are supplied by the user.