feat: assume data-setup will be a string
This commit removes the converter from the attribute setter/accessor and
uses it instead in the `firstUpdated()` lifecycle method. Why? Because
depending on which FE framework this component gets consumed, the value
of `dataSetup` will differ. It's almost always a `string`, but, in
React, it can actually come through as an object as well. So we have to
handle both possibilites. In which case, it's safest to assume we are
always recieving a string from the attribute value that we need to
convert to an object. In the case we do get an object, we just pass it
along to the `buildOptionsObj()` function.`