feat!: add type attribute and width fallbacks
This commit refactors `ix-video.ts` so that:

1. Less JSON parsing/object manipulation needs to be done to use
  dataSetup.
2. Fallback styling is more in line with user expectations and videojs
   defaults
3. There are no FOUC issues when using fallback styling
4. Since shadow-dom is off, we defensively add unique identifiers to
  video players

One of the main changes in this commit is the use of the
`connectedCallback()` lifecycle method to set a minimum width for the
host component and the video element.

Another significant change is  the use of the `type` attribute. Although
it defaults to HLS, adding it allows us to provide a way for users to
use this component with other video filetypes