feat: spread host attribute to video component
This commit spreads any attribute without an accessor already defined to
the `<video>` element. This allows us to "support" default `<video>`
attributes without the overhead of having to use a `@property` decorator
on each attribute.

This commit also refactor the options building logic into a private
instance method so that the `firstUpdated()` lifecycle hook is a little
easier to grok.