chore: remove hard-coded element id
This commmit removed the `generateUID` function and `uid` state value.
It also removes the inline `id` attriute being set on the `video`
element. This is because VJS will generate a unqiue `id` for us on the
fly. We're actually increasing the chances of a id collision by doing
this ourselves.