fix: ensure props get reflected on custom element
This commit fixes an issue where properties set on the element, when
read/set as Lit properties, were then no longer visible on the custom
element. This could lead to issues when the element was
mounted/unmounted several times in succession.

After this commit, the properties are reflected onto the custom element.
This means that even properties that the user left unset, IE default to
an empty string, will get relfect. A future commit should probably find
a way to keep the defaults from being reflected.