refactor: ensure unset attrs not set to emtpy str
Before this commit, if a user didn't set an attribute, it would be
"reflected" as an empty attribute rather than not bieng reflected at
all.

IE, `<ix-video source="<my-source">` would later be seen in the DOM as
`ix-video source="<my-source>" width height data-setup>` instead of just
`<ix-video source="my-source">`.

After this commit, the unset attributes are not then "reflected" as
emtpy strings.

This commit changes the `convertDataSetupStringToObject` helper function
to account for instances where `data-setup` could be unset.