refactor: use converter to change data-setup json string into obj
Before this commit, data-setup was set/read as a string and had to be
parsed, each time it was needed, into a POJO.

After this commit, data-setup is immediately converted to a POJO using a
custom converter. This allows us to assume throught the lifetime of the
component that data-setup will always be an object.