refactor: store options in state
Before this commit, options were being set in different places throught
the component, resulting in unpredictable state for the options value.

After this commit, options will be set by the contructor to match the
properties it's copying/modifying into state.

A subsequent commit will define getters/setters that ensure the shape of
`options` stays consistent and up-to-date with property changes.