chore: remove data-setup typing in state
The `dataSetup` in state cannot be cast as `DataSetup` type because the
compiler will think that `width/height` are defined on it despite them
being declared as optional. Then it will get angry when we try to spread
`DataSetup` with `this.options`. Annoying but only way to get the prod
bundle to build is to not cast it when first adding it to state.`