build: use vite to build bundle
Prior to this commit, rollup was being use to bundle the application
directly. This was 1) slow and 2) difficult to use when trying to also
support UMD modules. The CJS and ES dependencies that are used in this
project don't play nicely with the `commonjs()` rollup plugin, making
it difficult to create a UMD bundle. Using vite side-steps this issue.

The one major drawback here is that the CSS cannot be inlined into the
JS. So we have to both 1) provde an index.css file that includes the
videojs styles and 2) check if it's being used and fallback to using a
`<link>` tag that references the CSS if it's not.